+++ hc/front/tests/test_add_check.py
@@ -39,6 +39,12 @@ class AddCheckTestCase(BaseTestCase):
 
         self.assertRedirects(r, self.redirect_url)
 
+    def test_redirect_preserves_querystring(self):
+        referer = self.redirect_url + "?tag=foo"
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, self._payload(), HTTP_REFERER=referer)
+        self.assertRedirects(r, referer)
+
     def test_it_saves_cron_schedule(self):
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, self._payload(kind="cron"))