+++ hc/front/tests/test_update_timeout.py
@@ -31,6 +31,14 @@ class UpdateTimeoutTestCase(BaseTestCase):
         expected_aa = self.check.last_ping + td(seconds=3600 + 60)
         self.assertEqual(self.check.alert_after, expected_aa)
 
+    def test_redirect_preserves_querystring(self):
+        referer = self.redirect_url + "?tag=foo"
+        payload = {"kind": "simple", "timeout": 3600, "grace": 60}
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.post(self.url, data=payload, HTTP_REFERER=referer)
+        self.assertRedirects(r, referer)
+
     def test_it_does_not_update_status_to_up(self):
         self.check.last_ping = timezone.now() - td(days=2)
         self.check.status = "down"